에어마우스 파트: - 코드 최적화

+++ AirMouse_fw/Core/Inc/kalman_filter.h
@@ -10,16 +10,16 @@
 
 
 typedef struct {
-    float q;  // 프로세스 노이즈 공분산
-    float r;  // 측정 노이즈 공분산
-    float x;  // 값 추정치
-    float p;  // 오류 공분산 추정치
-    float k;  // 칼만 이득
+  double q;  // 프로세스 노이즈 공분산
+  double r;  // 측정 노이즈 공분산
+  double x;  // 값 추정치
+  double p;  // 오류 공분산 추정치
+  double k;  // 칼만 이득
 } KalmanFilter;
 
 
-void KalmanFilter_Init(KalmanFilter *kf, float process_noise, float measurement_noise, float initial_value);
-float KalmanFilter_Update(KalmanFilter *kf, float measurement);
+void KalmanFilter_Init(KalmanFilter *kf, double process_noise, double measurement_noise, double initial_value);
+double KalmanFilter_Update(KalmanFilter *kf, double measurement);
 
 
 #endif /* INC_KALMAN_FILTER_H_ */
